+++ Components/Core/Conditions/Categories/Spells.lua
@@ -211,48 +211,31 @@ ConditionCategory:RegisterCondition(3,	 "REACTIVE", {
 			ConditionObject:GenerateNormalEventString("SPELL_UPDATE_USABLE")
 	end,
 })
-
-
-ConditionCategory:RegisterCondition(3.5,  "OVERLAYED", {
-	text = L["CONDITIONPANEL_OVERLAYED"],
-	tooltip = L["CONDITIONPANEL_OVERLAYED_DESC"],
+ConditionCategory:RegisterCondition(3.1, "CURRENTSPELL", {
+	text = L["CONDITIONPANEL_CURRENTSPELL"],
+	tooltip = L["CONDITIONPANEL_CURRENTSPELL_DESC"],
 
 	bool = true,
 	
 	name = function(editbox)
-		editbox:SetTexts(L["CONDITIONPANEL_OVERLAYED"], L["CNDT_ONLYFIRST"])
+		editbox:SetTexts(L["SPELLTOCHECK"], L["CNDT_ONLYFIRST"])
 		editbox:SetLabel(L["SPELLTOCHECK"])
 	end,
 	useSUG = true,
 	unit = false,
-	icon = "Interface\\Icons\\inv_shield_05",
+	formatter = TMW.C.Formatter.BOOL_USABLEUNUSABLE,
+	icon = "Interface\\Icons\\ability_rogue_ambush",
 	tcoords = CNDT.COMMON.standardtcoords,
 	Env = {
-		IsSpellOverlayed = IsSpellOverlayed,
-		OverlayedNameMap = {}
+		IsCurrentSpell = IsCurrentSpell,
 	},
-	funcstr = function(c)
-		local module = CNDT:GetModule("OVERLAYED", true)
-		if not module then
-			module = CNDT:NewModule("OVERLAYED", "AceEvent-3.0")
-
-			local function handleEvent(event, arg1)
-				Env.OverlayedNameMap[strlowerCache[GetSpellInfo(arg1)]] = arg1
-			end
-
-			module:RegisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW", handleEvent)
-			module:RegisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_HIDE", handleEvent)
-		end
-
-		return [[BOOLCHECK( IsSpellOverlayed(OverlayedNameMap[c.NameFirst] or (isNumber[c.NameFirst] and c.NameFirst) or 0) )]]
-	end,
+	funcstr = [[BOOLCHECK( IsCurrentSpell(c.NameFirst) )]],
 	events = function(ConditionObject, c)
 		return
-			ConditionObject:GenerateNormalEventString("SPELL_UPDATE_USABLE"),
-			ConditionObject:GenerateNormalEventString("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW"),
-			ConditionObject:GenerateNormalEventString("SPELL_ACTIVATION_OVERLAY_GLOW_HIDE")
+			ConditionObject:GenerateNormalEventString("CURRENT_SPELL_CAST_CHANGED")
 	end,
 })
+
 
 ConditionCategory:RegisterCondition(4,	 "MANAUSABLE", {
 	text = L["CONDITIONPANEL_MANAUSABLE"],