+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed an issue with Unit Conditions where the initial state of the conditions sometimes wouldn't be taken into account.
 * Changed the Slowed equivalency to track Crippling Poison by ID to prevent it from picking up the Rogue buff by the same name.
 * When scrolling with the mousewheel, sliders that happen to land under your mouse will no longer be adjusted as long as your cursor does not move.
+* Fixed an issue where the Artificial Maximum setting for Bar groups was not properly saving its value as a number.
 
 ## v8.6.8
 * Re-releasing TellMeWhen for Retail WoW as 8.6.8 so it will be the latest file for people with out-of-date Twitch apps.